+++ tests/test_filterbank.py
@@ -2,11 +2,12 @@ import unittest
 import numpy as np
 
 import audiotools as audio
-import audiotools.filter.filterbank as bank
+import audiotools.filter.bank.filterbank as bank
 from audiotools.filter import gammatone_filt
 from audiotools.filter import butterworth_filt
 import numpy.testing as testing
-from audiotools.filter import create_filterbank
+from audiotools.filter.bank import create_filterbank
+from audiotools.filter.bank import auditory_gamma_bank, octave_bank
 
 
 class test_oaudio(unittest.TestCase):
@@ -50,7 +51,6 @@ class test_oaudio(unittest.TestCase):
         # Amplitudes should be 0.5 (two sided spectrum)
         assert np.all((amps - 0.5) <= 0.01)
 
-
     def test_gammatone(self):
         # Test if filter coefficents are equivalent tho those of
         # individual filters
@@ -75,7 +75,7 @@ class test_oaudio(unittest.TestCase):
         # create filterbank an run signal
         bank = create_filterbank(fc, bw, 'gammatone', 48000)
         sig_out = bank.filt(sig)
-        #check amplitudes at fc of every filter
+        # check amplitudes at fc of every filter
         amps = np.zeros(len(fc))
         for i_fc, f in enumerate(fc):
             spec = sig_out.ch[i_fc].to_freqdomain()
@@ -96,8 +96,6 @@ class test_oaudio(unittest.TestCase):
         out = gamma.filt(sig)
         assert(out.shape[1:] == (2, 1, 2))
 
-
-
     def test_brickwall(self):
         # Test the gain at fc
         sig = audio.Signal(1, 1, 48000)
@@ -109,14 +107,13 @@ class test_oaudio(unittest.TestCase):
         # create filterbank an run signal
         bank = create_filterbank(fc, bw, 'brickwall', 48000)
         sig_out = bank.filt(sig)
-        #check amplitudes at fc of every filter
+        # check amplitudes at fc of every filter
         amps = np.zeros(len(fc))
         for i_fc, f in enumerate(fc):
             spec = sig_out.ch[i_fc].to_freqdomain()
             amps[i_fc] = np.abs(spec[spec.freq==f])
         # Amplitudes hould be 0.5 (double sided spectrum)
         assert np.all((amps - 0.5) <= 0.01)
-
 
     def test_set_params(self):
         fc = [100, 200, 5000]
@@ -132,3 +129,32 @@ class test_oaudio(unittest.TestCase):
             a_bank = gamma.coefficents[2:, i_filt]
             testing.assert_array_equal(a, a_bank)
             testing.assert_array_equal(b, b_bank)
+
+    def test_auditory_gamma_bank(self):
+        filt_bank = auditory_gamma_bank(fs = 48000)
+        assert isinstance(filt_bank, bank.GammaToneBank)
+
+        fcs = audio.freqarange(16, 16000, 1, 'erb')
+        testing.assert_array_equal(filt_bank.fc, fcs)
+
+    def test_butterworth_zero(self):
+        # Catch a fixed bug where filters would oscillate due to wrong inital
+        # states
+        fs = 48000
+        sig = audio.Signal(1, 1, fs)
+        fb = create_filterbank([125, 500, 1000],
+                               [50, 55, 58],
+                               'butter', fs,
+                               order=4)
+        out = fb.filt(sig)
+        testing.assert_allclose(out, 0)
+
+    def test_default_octave_bank(self):
+        # Test that power in all channels is aproximatly equal when applyting to
+        # white noise
+        fs = 48000
+        filt_bank = octave_bank(fs)
+        noise = audio.Signal(1, 10, 48000).add_noise('pink')
+        bank_out = filt_bank.filt(noise)
+        power = np.var(bank_out, axis=0)
+        assert power.std() < 0.01